Create actions file and refactor main.rb file
